requirements.txt
fastapi >=0.63,<0.64.0
uvicorn
python-dotenv
python-multipart
requests